+++ src/components/Multiselectableautocomplete.tsx
@@ -11,10 +11,10 @@ import { Input } from 'src/components/Input';
 import { Label } from 'src/components/Label';
 import { useHasClickedOutside } from 'src/hooks/useHasClickedOutside';
 import classes from 'src/styles/multiselectableAutocomplete.module.css';
-import { MultiselectableAutocompleteProps } from 'src/types/components/MultiselectableAutocomplete';
+import { MultiselectableAutocompleteProps } from 'src/types/components/Multiselectableautocomplete';
 import { classNames } from 'src/utils/css/classNames';
 
-export function MultiselectableAutocomplete({
+export function Multiselectableautocomplete({
     label,
     name,
     placeholder,
