+++ src/components/main.ts
@@ -16,7 +16,7 @@ export * from 'src/components/DropDownItem';
 export * from 'src/components/Select';
 export * from 'src/components/Input';
 export * from 'src/components/VirtualizedList';
-export * from 'src/components/MultiselectableAutocomplete';
+export * from 'src/components/Multiselectableautocomplete';
 export * from 'src/components/Table';
 export * from 'src/components/Dialog';
 export * from 'src/components/DialogActions';
